Fix parser and ps1
Fix bgm id in parser
Fix output encoding in ps1
Add option to keep json in ps1

+++ utils/Process-ScoresFolder.ps1
@@ -1,12 +1,14 @@
 param (
     [string]$Folder = $(Read-Host "Root folder: "),
     [string]$MusicDb = $(Read-Host "Music db: "),
-    [string]$Visualizer = $(Read-Host "Visualizer cmd: ")
+    [string]$Visualizer = $(Read-Host "Visualizer cmd: "),
+    [bool]$KeepJson = $False
 )
 
-$Folder = "D:\Develop\Temporary Data\bang\musicscore"
-$MusicDb = "D:\Develop\Temporary Data\bang\musics.json"
-$Visualizer = "D:\Develop\Toys\BandoriScore\BandoriScoreVisualizer\bin\Release\BandoriScoreVisualizer.exe"
+$PSDefaultParameterValues['Out-File:Encoding'] = 'utf8'
+#$Folder = "D:\Develop\Temporary Data\bang\musicscore"
+#$MusicDb = "D:\Develop\Temporary Data\bang\musics.json"
+#$Visualizer = "D:\Develop\Toys\BandoriScore\BandoriScoreVisualizer\bin\Release\BandoriScoreVisualizer.exe"
 
 cd $PSScriptRoot
 
@@ -15,17 +17,23 @@ $regex = New-Object System.Text.RegularExpressions.Regex -ArgumentList "([^_]+)\
 
 ForEach ($di in $root.GetDirectories()) {
     ForEach ($fi in $di.GetFiles("*.txt")) {
-        Write-Host $fi.Name
+        Write-Host $fi.Directory.Name $fi.Name
         $name = $fi.Name;
         $diff = $regex.Match($name).Groups[1].Value
         if ($diff -eq "command") {
             continue
         }
 
+        $fullName = $fi.FullName
         $jsonName = $fi.FullName.Replace(".txt", ".json")
         $pngName = $fi.FullName.Replace(".txt", ".png")
-        Get-Content $fi.FullName | node.exe ..\parser.js -m $MusicDb -d $diff > $jsonName
+        #fuck powershell encoding
+        #Get-Content $fi.FullName | node.exe ..\parser.js -m $MusicDb -d $diff > $jsonName
+        cmd /c "node ..\parser.js -m `"$MusicDb`" -d $diff < `"$fullName`" > `"$jsonName`""
+
         & $Visualizer $jsonName $pngName
-        Remove-Item $jsonName
+        if (-Not $KeepJson) {
+            Remove-Item $jsonName
+        }
     }
 }
